edit dialog

+++ src/app/blog-dashboard/delete-dialog/delete-dialog.component.ts
@@ -24,7 +24,9 @@ export class DeleteDialogComponent implements OnInit {
   }
 
   async deleteConfirmed() {
-    await this.blogService.delete(this.post.id);
+    if (this.post.id !== 'new') {
+      await this.blogService.delete(this.post.id);
+    }
     this.dialogRef.close();
   }
 }
